% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-interfaces.R
\docType{data}
\name{CloudInterface}
\alias{CloudInterface}
\title{Class of objects for interacting with objects in a cloud storage system.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CloudInterface
}
\arguments{
\item{get_fn}{function. Used to get an object from the cloud.}

\item{put_fn}{function. Used to put an object to the cloud.}

\item{exists_fn}{function. Used to check if an object exists in the cloud.}
}
\description{
A CloudInterface object holds the methods to read, write, and check
for existence of objects in some external resource. We assume that
objects are stored in some kind of key/value structure with one-to-one
correspondence. We also assume that you are using disk space as an intermediary
between the cloud storage system and your R session.
}
\keyword{datasets}

