% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_write.R
\name{csmpi_custom_write}
\alias{csmpi_custom_write}
\title{Write to the cloud using custom cloud and disk interfaces.}
\usage{
csmpi_custom_write(obj, key, cloud_interface, disk_interface, params,
  use_disk_cache = getOption("csmpi.use_disk_cache", FALSE),
  num_retries = getOption("csmpi.num_retries", 3),
  overwrite_disk_cache = FALSE,
  overwrite_cloud_object = getOption("csmpi.force_write", TRUE), cloud_name_,
  storage_format_)
}
\arguments{
\item{obj}{ANY. The object to write to the cloud.}

\item{key}{simple string. The key to record the object with.}

\item{cloud_interface}{CloudInterface object. The cloud interface to use.}

\item{disk_interface}{DiskInterface object. The disk interface to use.}

\item{params}{list. Additional arguments to pass to interface methods.}

\item{use_disk_cache}{logical. Whether or not to cache the object on disk.
Defaults to the option \code{csmpi.use_disk_cache}, or \code{FALSE} if
the option is not set.}

\item{num_retries}{numeric. How many attempts to write before raising an error.
Defaults to the option \code{csmpi.num_retries}, or \code{3} if
the option is not set.}

\item{overwrite_disk_cache}{logical. If the object is already cached on disk,
do we want to overwrite the cached copy?}

\item{cloud_name_}{simple_string. The name of the cloud interface, used to
create unique file names for caching. If not provided, will be computed
from the cloud interface using non-standard evaluation.}

\item{storage_format_}{simple_string. The name of the disk interface, used to
create unique file names for caching. If not provided, will be computed
from the disk interface using non-standard evaluation.}
}
\description{
Write to the cloud using custom cloud and disk interfaces.
}
