% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_exists.R
\name{csmpi_exists}
\alias{csmpi_exists}
\title{Check if an object exits in the cloud.}
\usage{
csmpi_exists(key, cloud_name, params)
}
\arguments{
\item{key}{simple string. The key under which the object is stored.}

\item{cloud_name}{simple string. The name of the cloud interface to use.}

\item{params}{list. Additional params for the cloud interface's \code{exists}
method.}
}
\description{
Check if an object exits in the cloud.
}
