% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_read.R
\name{csmpi_read}
\alias{csmpi_read}
\title{Read an R object to the cloud.}
\usage{
csmpi_read(key, cloud_name, storage_format, params,
  use_session_cache = getOption("csmpi.use_session_cache", TRUE),
  use_disk_cache = getOption("csmpi.use_disk_cache", FALSE),
  num_retries = getOption("csmpi.num_retries", 3))
}
\arguments{
\item{key}{simple string. The key to record the object with.}

\item{cloud_name}{simple string. The name of the cloud interface to use.}

\item{storage_format}{simple_string. The name of the disk interface to use.}

\item{use_session_cache}{logical. Whether or not to cache the object in memory.
Defaults to the option \code{csmpi.use_session_cache}, or \code{FALSE} if
the option is not set.}

\item{use_disk_cache}{logical. Whether or not to cache the object on disk.
Defaults to the option \code{csmpi.use_disk_cache}, or \code{FALSE} if
the option is not set.}

\item{num_retries}{numeric. How many attempts to write before raising an error.
Defaults to the option \code{csmpi.num_retries}, or \code{3} if
the option is not set.}
}
\description{
Read an R object to the cloud.
}

