% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_write.R
\name{csmpi_write}
\alias{csmpi_write}
\title{Write an R object to the cloud.}
\usage{
csmpi_write(obj, key, cloud_name, storage_format, params,
  use_disk_cache = getOption("csmpi.use_disk_cache", FALSE),
  num_retries = getOption("csmpi.num_retries", 3),
  overwrite_disk_cache = FALSE)
}
\arguments{
\item{obj}{ANY. The object to write to the cloud.}

\item{key}{simple string. The key to record the object with.}

\item{cloud_name}{simple string. The name of the cloud interface to use.}

\item{storage_format}{simple_string. The name of the disk interface to use.}

\item{params}{list. Additional arguments to pass to interface methods.}

\item{use_disk_cache}{logical. Whether or not to cache the object on disk.
Defaults to the option \code{csmpi.use_disk_cache}, or \code{FALSE} if
the option is not set.}

\item{num_retries}{numeric. How many attempts to write before raising an error.
Defaults to the option \code{csmpi.num_retries}, or \code{3} if
the option is not set.}

\item{overwrite_disk_cache}{logical. If the object is already cached on disk,
do we want to overwrite the cached copy?}
}
\description{
Write an R object to the cloud.
}

