% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-interfaces.R
\docType{data}
\name{DiskInterface}
\alias{DiskInterface}
\title{Class of objects for interacting with objects downloaded to disk from the cloud.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DiskInterface
}
\arguments{
\item{read_fn}{function. Used to read objects from disk to R.}

\item{write_fn}{function. Used to write objects from R to disk.}
}
\description{
A DiskInterface objects holds methods for reading and writing objects from disk in
a particular format.  We assume that your "disk" is your local disk, as we use
base R functions such as \code{file.exists} to interact with objects.
}
\keyword{datasets}

