% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_exists.R
\name{exists}
\alias{exists}
\title{Check if an object exits in the cloud with a custom cloud interface.}
\usage{
exists(key, cloud_interface, params)
}
\arguments{
\item{key}{simple string. The key under which the object is stored.}

\item{cloud_interface}{CloudInterface object. The cloud interface to use.}

\item{params}{list. Additional params for the cloud interface's \code{exists}
method.}
}
\description{
Check if an object exits in the cloud with a custom cloud interface.
}

