% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_read.R
\name{read}
\alias{read}
\title{Read from the cloud using custom cloud and disk interfaces.}
\usage{
read(key, cloud_interface, disk_interface, params,
  use_session_cache = getOption("csmpi.use_session_cache", TRUE),
  use_disk_cache = getOption("csmpi.use_disk_cache", FALSE),
  num_retries = getOption("csmpi.num_retries", 3), session_cache_key,
  disk_cache_filename, cloud_name_, storage_format_)
}
\arguments{
\item{cloud_interface}{CloudInterface object. The cloud interface to use.}

\item{disk_interface}{DiskInterface object. The disk interface to use.}

\item{use_session_cache}{logical. Whether or not to cache the object in memory.
Defaults to the option \code{csmpi.use_session_cache}, or \code{FALSE} if
the option is not set.}

\item{cloud_name_}{simple_string. The name of the cloud interface, used to
create unique file names for caching. If not provided, will be computed
from the cloud interface using non-standard evaluation.}

\item{storage_format_}{simple_string. The name of the disk interface, used to
create unique file names for caching. If not provided, will be computed
from the disk interface using non-standard evaluation.}
}
\description{
Read from the cloud using custom cloud and disk interfaces.
}

