% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_read.R
\name{read_}
\alias{read_}
\title{Non-NSE version of \code{read}.}
\usage{
read_(key, cloud_interface, disk_interface, params,
  use_session_cache = getOption("csmpi.use_session_cache", FALSE),
  use_disk_cache = getOption("csmpi.use_disk_cache", FALSE),
  num_retries = getOption("csmpi.num_retries", 3), session_cache_key,
  disk_cache_filename, cloud_name_, storage_format_)
}
\arguments{
\item{key}{simple string. The key to record the object with.}

\item{cloud_interface}{CloudInterface object. The cloud interface to use.}

\item{disk_interface}{DiskInterface object. The disk interface to use.}

\item{params}{list. Additional arguments to pass to interface methods.}

\item{use_session_cache}{logical. Whether or not to cache the object in memory.
Defaults to the option \code{csmpi.use_session_cache}, or \code{FALSE} if
the option is not set.}

\item{use_disk_cache}{logical. Whether or not to cache the object on disk.
Defaults to the option \code{csmpi.use_disk_cache}, or \code{FALSE} if
the option is not set.}

\item{num_retries}{numeric. How many attempts to write before raising an error.
Defaults to the option \code{csmpi.num_retries}, or \code{3} if
the option is not set.}

\item{cloud_name_}{simple_string. The name of the cloud interface, used to
create unique file names for caching. If not provided, will be computed
from the cloud interface using non-standard evaluation.}

\item{storage_format_}{simple_string. The name of the disk interface, used to
create unique file names for caching. If not provided, will be computed
from the disk interface using non-standard evaluation.}
}
\description{
Non-NSE version of \code{read}.
}

