% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cmd.R, R/s3cmd_helpers.R
\name{s3cmd}
\alias{s3cmd}
\alias{get_using_s3cmd}
\alias{put_using_s3cmd}
\alias{exists_using_s3cmd}
\alias{which_s3cmd}
\alias{options_for_s3cmd_get}
\alias{options_for_s3cmd_put}
\title{Tooling for s3cmd interface.}
\usage{
get_using_s3cmd(key, filename, params)

put_using_s3cmd(key, filename, params)

exists_using_s3cmd(key, params)

which_s3cmd()

options_for_s3cmd_get(params)

options_for_s3cmd_put(params)
}
\arguments{
\item{key}{simple string. The key to read/write to in S3.}

\item{filename}{simple string. The file to read/write to on
disk when reading/writing from S3.}

\item{params}{list. Addiitonal parameters to pass to read/write functions.}
}
\description{
Tooling for s3cmd interface.
}
