% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-s3cmd_shortcuts.R
\name{s3cmdexists}
\alias{s3cmdexists}
\title{Check if an object exists in S3 using s3cmd.}
\usage{
s3cmdexists(name, path = s3cmd_default_path(), ...)
}
\arguments{
\item{name}{simple string. The key used to store the object.}

\item{path}{simple string. The s3 bucket where the object is stored.}

\item{...}{additional parameters to pass to interface objects via \code{params}.}
}
\description{
This function will check if there is an object in AWS S3 with the given key.
It will not check the integrity of the object - it will use s3cmd to check
if there is an object with that name.
}

