% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-s3cmd_shortcuts.R
\name{s3cmdread}
\alias{s3cmdread}
\title{Read an object from S3 into an R session using s3cmd.}
\usage{
s3cmdread(name, path = s3cmd_default_path(), storage_format = "RDS", ...)
}
\arguments{
\item{name}{simple string. The key used to store the object.}

\item{path}{simple string. The s3 bucket where the object is stored.}

\item{storage_format}{simple string. The name of the disk interface in
/code{DEFAULT_DISK_INTERFACES} to use to read the object from disk into
the R session.}

\item{...}{additional parameters to pass to interface objects via \code{params}.}
}
\description{
This function will read an object from AWS S3 using
s3cmd, download it to disk, and read it into the current
R session.
}
\details{
To use this effectively, make sure that you are reading in
objects in the same format as you write them! (As of version 0.0.8,
the only storage format is RDS, but expect there to be more of them later)
}

