% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-s3cmd_shortcuts.R
\name{s3cmdstore}
\alias{s3cmdstore}
\title{Store an object from an R session to S3 using s3cmd.}
\usage{
s3cmdstore(obj, name, path = s3cmd_default_path(), storage_format = "RDS",
  ..., force = FALSE)
}
\arguments{
\item{obj}{ANY. The R object to store.}

\item{name}{simple string. The key used to store the object.}

\item{path}{simple string. The s3 bucket where the object is stored.}

\item{...}{additional parameters to pass to interface objects via \code{params}.}
}
\description{
This function will write an object to a file on disk, and upload
the file to AWS S3 using s3cmd.
}
\details{
To use this effectively, make sure that you are reading in
objects in the same format as you write them! (As of version 0.0.8,
the only storage format is RDS, but expect there to be more of them later)
}

