% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnostic_write.R
\name{write_}
\alias{write_}
\title{Non-NSE version of \code{write}.}
\usage{
write_(obj, key, cloud_interface, disk_interface, params,
  use_disk_cache = getOption("csmpi.use_disk_cache", FALSE),
  num_retries = getOption("csmpi.num_retries", 3),
  overwrite_disk_cache = FALSE, cloud_name_, storage_format_)
}
\arguments{
\item{cloud_interface}{CloudInterface object. The cloud interface to use.}

\item{disk_interface}{DiskInterface object. The disk interface to use.}

\item{cloud_name_}{simple_string. The name of the cloud interface, used to
create unique file names for caching. If not provided, will be computed
from the cloud interface using non-standard evaluation.}

\item{storage_format_}{simple_string. The name of the disk interface, used to
create unique file names for caching. If not provided, will be computed
from the disk interface using non-standard evaluation.}
}
\description{
Non-NSE version of \code{write}.
}

