% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cmd.R
\name{s3cmd}
\alias{s3cmd}
\alias{which_s3cmd}
\alias{extract_bucket_location}
\alias{s3cmd_default_path}
\alias{get_using_s3cmd}
\alias{put_using_s3cmd}
\alias{exists_using_s3cmd}
\title{Tooling for s3cmd interface.}
\usage{
which_s3cmd()

extract_bucket_location(params)

s3cmd_default_path()

get_using_s3cmd(key, filename, params)

put_using_s3cmd(key, filename, params)

exists_using_s3cmd(key, params)
}
\arguments{
\item{params}{list. Addiitonal parameters to pass to read/write functions.}

\item{key}{simple string. The key to read/write to in S3.}

\item{filename}{simple string. The file to read/write to on
disk when reading/writing from S3.}
}
\description{
Tooling for s3cmd interface.
}
